FMCOMMS2 Sink: Use integer types for some parameters

+++ include/gnuradio/iio/fmcomms2_sink.h
@@ -36,10 +36,10 @@ namespace gr {
      public:
       typedef boost::shared_ptr<fmcomms2_sink> sptr;
 
-      static sptr make(const std::string &host,
-		    double frequency, double samplerate, double bandwidth,
+      static sptr make(const std::string &host, unsigned long long frequency,
+		    unsigned long samplerate, unsigned long bandwidth,
 		    bool ch1_en, bool ch2_en, bool ch3_en, bool ch4_en,
-		    unsigned int buffer_size, bool cyclic,
+		    unsigned long buffer_size, bool cyclic,
 		    const char *rf_port_select,
 		    double attenuation1, double attenuation2);
     };
