+++ lib/fmcomms2_sink_impl.h
@@ -41,10 +41,10 @@ namespace gr {
 		      bool ch1_en, bool ch2_en, bool ch3_en, bool ch4_en);
 
      public:
-      fmcomms2_sink_impl(const std::string &host,
-		    double frequency, double samplerate, double bandwidth,
+      fmcomms2_sink_impl(const std::string &host, unsigned long long frequency,
+		    unsigned long samplerate, unsigned long bandwidth,
 		    bool ch1_en, bool ch2_en, bool ch3_en, bool ch4_en,
-		    unsigned int buffer_size, bool cyclic,
+		    unsigned long buffer_size, bool cyclic,
 		    const char *rf_port_select,
 		    double attenuation1, double attenuation2);
